fix : block in byte

+++ src/fr/math/minecraft/client/world/Chunk.java
@@ -6,7 +6,7 @@ import org.joml.Vector3i;
 public class Chunk {
 
     private final Vector3i position;
-    private final int[] blocks;
+    private final byte[] blocks;
     private boolean empty;
 
     public final static int SIZE = 16;
@@ -17,7 +17,7 @@ public class Chunk {
 
     public Chunk(int x, int y, int z) {
         this.position = new Vector3i(x, y, z);
-        this.blocks = new int[VOLUME];
+        this.blocks = new byte[VOLUME];
         for (int blockX = 0; blockX < Chunk.SIZE; blockX++) {
             for (int blockY = 0; blockY < Chunk.SIZE; blockY++) {
                 for (int blockZ = 0; blockZ < Chunk.SIZE; blockZ++) {
@@ -29,15 +29,15 @@ public class Chunk {
         this.chunkMesh = null;
     }
 
-    public int[] getBlocks() {
+    public byte[] getBlocks() {
         return blocks;
     }
 
-    public void setBlock(int x, int y, int z, int block) {
+    public void setBlock(int x, int y, int z, byte block) {
         blocks[x + y * AREA + z * SIZE] = block;
     }
 
-    public int getBlock(int x, int y, int z) {
+    public byte getBlock(int x, int y, int z) {
         return blocks[x + y * AREA + z * SIZE];
     }
